backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
Pillow==10.0.0
python-docx==0.8.11
python-magic==0.4.27
Wand==0.6.13